Upgrade goldberg to normal mesh
